+++ src/tests/testEventLoop.cc
@@ -9,7 +9,7 @@
 #include "squid.h"
 #include "async/AsyncEngine.h"
 #include "compat/cppunit.h"
-#include "EventLoop.h"
+#include "async/EventLoop.h"
 #include "time/Engine.h"
 #include "unitTestMain.h"
