+++ src/tests/testStoreSupport.cc
@@ -8,7 +8,7 @@
 
 #include "squid.h"
 #include "async/Event.h"
-#include "EventLoop.h"
+#include "async/EventLoop.h"
 #include "testStoreSupport.h"
 
 /* construct a stock loop with event dispatching, a time service that advances
